% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{ADF}
\alias{ADF}
\title{Calculate Average Daily Flow (ADF) from Peak Hour Flow (PHF)}
\usage{
ADF(GPM)
}
\arguments{
\item{GPM}{Peak Hour Flow (PHF), in GPM}
}
\value{
Average Daily Flow (ADF), in MGD, using 10-state peaking factor method
}
\description{
Calculate Average Daily Flow (ADF) from Peak Hour Flow (PHF)
}
\examples{
ADF(1000)
}
