% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.Flow}
\alias{Draw.Flow}
\title{Draw Design Flow/Head lines (ICM Style)}
\usage{
Draw.Flow(pump, Q)
}
\arguments{
\item{pump}{pump object from New.Pumps()}

\item{Q}{Target Flow, in GPM}
}
\value{
Plot
}
\description{
Draw Design Flow/Head lines (ICM Style)
}
\examples{
Draw.Graph(500,100,25, 150,50,10)
pump <- New.Pump(c(0, 100,50,80, 100,60,150,40,200,20))
Draw.Pump(pump, 2)
Draw.Flow(pump,125)
}
