% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.Label.Point}
\alias{Draw.Label.Point}
\title{Draw Duty Point to Graph with label}
\usage{
Draw.Label.Point(Q, H, color = "red", DPLabel = "Capacity")
}
\arguments{
\item{Q}{Flow, GPM}

\item{H}{Head, FT}

\item{color}{optional}

\item{DPLabel}{optional}
}
\value{
Plot
}
\description{
Draw Duty Point to Graph with label
}
\examples{
Draw.Graph()
Draw.Label.Point(3000,50)
}
