% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.Legend}
\alias{Draw.Legend}
\title{Draw Legend}
\usage{
Draw.Legend(position = "bottomleft", LineText, LineColor, LineType,
  PointSymbol = rep(NA, length(LineType)), LineWeight = rep(2,
  length(LineType)))
}
\arguments{
\item{position}{location on graph, such as "top", "bottomleft", etc...}

\item{LineText}{Vector of text to be displayed}

\item{LineColor}{Vector of Line Colors}

\item{LineType}{Vector of Line Types}

\item{PointSymbol}{Vector of symbols (see pch)}

\item{LineWeight}{vector of Line weights}
}
\value{
Plot
}
\description{
Draw Legend
}
\examples{
Draw.Graph()
Draw.Legend("top", c("Field", "Model"), c("black", "blue"), c(1,2))
}
