% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.Point}
\alias{Draw.Point}
\title{Draw duty point to graph}
\usage{
Draw.Point(Q, H, color = "red")
}
\arguments{
\item{Q}{Flow, GPM}

\item{H}{Head, FT}

\item{color}{optional}
}
\value{
Plot
}
\description{
Draw duty point to graph
}
\examples{
Draw.Graph()
Draw.Point(3000,50)
}
