% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.System.Curve}
\alias{Draw.System.Curve}
\title{Draws a system curve}
\usage{
Draw.System.Curve(k, S, linecolor = "red", linetype = 1)
}
\arguments{
\item{k}{hydraulic gradient}

\item{S}{static head}

\item{linecolor}{optional}

\item{linetype}{optional}
}
\value{
Plot
}
\description{
Draws a system curve
}
\examples{
Draw.Graph()
Draw.System.Curve(5000*ke(6,120),10)
}
