% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.System.Envelope}
\alias{Draw.System.Envelope}
\title{Draw Enevlope of min.max system curve}
\usage{
Draw.System.Envelope(sc, linetype = 1, color = JEA.Grey,
  max.Q = 5000)
}
\arguments{
\item{sc}{System Curve Object from Get.System.Curve()}

\item{linetype}{optional}

\item{color}{optional}

\item{max.Q}{optional}
}
\value{
Plot
}
\description{
Draw Enevlope of min.max system curve
}
\examples{
Draw.Graph(500,100,25, 150,50,10)
sc <- matrix(c(2.5E-4,5,46.2e-4,25), nrow=2, byrow=TRUE)
Draw.System.Envelope(sc)
}
