% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Draw.VFD}
\alias{Draw.VFD}
\title{Draw Pump with VFD envelope}
\usage{
Draw.VFD(PumpCurve, numPumps = 1, color = JEA.Blue, linetype = 1)
}
\arguments{
\item{PumpCurve}{pump object from New.Pump()}

\item{numPumps}{Number of parallel pumps to draw}

\item{color}{optional}

\item{linetype}{optional}
}
\value{
Plot
}
\description{
Draw Pump with VFD envelope
}
\examples{
Draw.Graph(500,100,25, 150,50,10)
pump <- New.Pump(c(
  0, 100,
 50,  80,
100,  60,
150,  40,
200,  20
))
Draw.VFD(pump, 2)
}
