% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Get.Flow}
\alias{Get.Flow}
\title{Get Flow from Head}
\usage{
Get.Flow(PumpCurve, H)
}
\arguments{
\item{PumpCurve}{pump object from New.Pump()}

\item{H}{Head, in FT}
}
\value{
Flow, in GPM
}
\description{
Get Flow from Head
}
\examples{
pump <- New.Pump(c(0, 100,50,80, 100,60,150,40,200,20))
Get.Head(pump,75)
}
