% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Get.System.Curve}
\alias{Get.System.Curve}
\title{Calc System Curve}
\usage{
Get.System.Curve(Q, TDH, min = 0.05, max = 0.95, min.flow = 10)
}
\arguments{
\item{Q}{Vector of flows}

\item{TDH}{Vectors of heads}

\item{min}{optional, minima quantile}

\item{max}{optional, minima quantile}

\item{min.flow}{min.flow for dynamic headloss calculation}
}
\value{
system curve object
}
\description{
Calc System Curve
}
\examples{
data <- data.frame(Q=c(rnorm(abs(rnorm(100,0))),rnorm(100,500)), H=c(rnorm(100,20), rnorm(100,100)))
Get.System.Curve(data$Q, data$H)
}
