% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{Import.HQ}
\alias{Import.HQ}
\title{Import Head/Flow from Model CSV}
\usage{
Import.HQ(filename)
}
\arguments{
\item{filename}{string, file path and name of CSV}
}
\value{
data.frame with two columns Q (Flow) and H (Head).
}
\description{
Import Head/Flow from Model CSV
}
