% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{ke}
\alias{ke}
\title{calculates the unit hydraulic gradient}
\usage{
ke(D, C)
}
\arguments{
\item{D}{Pipe diameter, in inches}

\item{C}{Hazen-Williams friction factor, unitless}
}
\value{
the hydraulic gradient per linear foot.
}
\description{
calculates the unit hydraulic gradient
}
\examples{
# Calculate the k for 1000 LF of 12-inch diameter pipe with C = 120
1000*ke(12,120)
}
