% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{MAF}
\alias{MAF}
\title{Maximum Allowable Flow}
\usage{
MAF(SC, MaxPSI = 60)
}
\arguments{
\item{SC}{System curve object}

\item{MaxPSI}{maximum allowable Pressure}
}
\value{
Flow, in GPM
}
\description{
Maximum Allowable Flow
}
\examples{
data <- data.frame(Q=c(rnorm(abs(rnorm(100,0))),rnorm(100,500)), H=c(rnorm(100,20), rnorm(100,100)))
Get.System.Curve(data$Q, data$H)
}
