% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{New.Pump}
\alias{New.Pump}
\title{New Pump Object}
\usage{
New.Pump(QH, n = 1)
}
\arguments{
\item{QH}{vector of flow and head by row}

\item{n}{number of pumps}
}
\value{
Pump Flow, in GPM
}
\description{
New Pump Object
}
\examples{
New.Pump(c(
  0, 100,
 50,  80,
100,  60,
150,  40,
200,  20
))
}
