% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{PF}
\alias{PF}
\title{Calculates the peaking factor for a given ADF, in GPD, using the 10-state peaking factor method}
\usage{
PF(GPD)
}
\arguments{
\item{GPD}{Average Daily Flow (ADF), in GPD}
}
\value{
peaking factor, unitless
}
\description{
Calculates the peaking factor for a given ADF, in GPD, using the 10-state peaking factor method
}
\examples{
PF(1e6)
}
