% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{PHF}
\alias{PHF}
\title{Peak Hour Flow (PHF)}
\usage{
PHF(GPD)
}
\arguments{
\item{GPD}{Average Daily Flow (ADF), in GPD}
}
\value{
Peak Hour Flow (PHF), in GPM, using the 10-state peaking factor method
}
\description{
Peak Hour Flow (PHF)
}
\examples{
PHF(1e6)
}
