% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{v}
\alias{v}
\title{Calculate pipe velocity}
\usage{
v(Q_gpm, D_inches)
}
\arguments{
\item{Q_gpm}{Pipe flow, in GPM}

\item{D_inches}{Pipe diameter, in inches}
}
\value{
pipe velocity, in FPS
}
\description{
Calculate pipe velocity
}
\examples{
v(1000,12)
}
