% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{VFD}
\alias{VFD}
\title{Adjust pump curve for VFD speed}
\usage{
VFD(pumpc = pump, Hz)
}
\arguments{
\item{pumpc}{pump object from New.Pumps}

\item{Hz}{VFD Speed, in Hz}
}
\value{
Pump Curve
}
\description{
Adjust pump curve for VFD speed
}
\examples{
#'
pump <- New.Pump(c(0, 100,50,80, 100,60,150,40,200,20))
VFD(pump,45)
}
