% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpR.R
\name{ADF}
\alias{ADF}
\title{Calculate Average Daily Flow (ADF) from Peak Hour Flow (PHF)}
\usage{
ADF(GPM)
}
\arguments{
\item{GPM}{peak hour flow (PHF) in gpm}
}
\value{
average daily flow (ADF) in mgd
}
\description{
Calculate Average Daily Flow (ADF) from Peak Hour Flow (PHF)
}
\examples{
ADF(1000)
}
